requirements.txt
streamlit>=1.30.0
pandas>=2.1.0
numpy>=1.24.0
plotly>=5.18.0
duckdb>=0.9.0
lightgbm>=4.1.0
scikit-learn>=1.3.0
google-generativeai>=0.8.0
llama-cpp-python>=0.3.0
huggingface-hub>=0.25.0
python-dotenv>=1.0.0
pyarrow>=14.0.0
